# transmute: Mark edges by byte sets, not byte values